requirements.txt
Django==1.11.7
django-crispy-forms==1.7.0
pytz==2017.3
six==1.11.0
